Fix anti dump x64 compatibility

+++ Confuser.Runtime/AntiDump.cs
@@ -70,7 +70,7 @@ namespace Confuser.Runtime {
 				*(uint*)mdHdr = 0;
 				mdHdr += 12;
 				mdHdr += *(uint*)mdHdr;
-				mdHdr = (byte*)(((uint)mdHdr + 7) & ~3);
+				mdHdr = (byte*)(((ulong)mdHdr + 7) & ~3UL);
 				mdHdr += 2;
 				ushort numOfStream = *mdHdr;
 				mdHdr += 2;
@@ -198,7 +198,7 @@ namespace Confuser.Runtime {
 				*(uint*)mdHdrPtr = 0;
 				mdHdrPtr += 12;
 				mdHdrPtr += *(uint*)mdHdrPtr;
-				mdHdrPtr = (byte*)(((uint)mdHdrPtr + 7) & ~3);
+				mdHdrPtr = (byte*)(((ulong)mdHdrPtr + 7) & ~3UL);
 				mdHdrPtr += 2;
 				ushort numOfStream = *mdHdrPtr;
 				mdHdrPtr += 2;
